requirements.txt
absl-py==0.8.1
astor==0.8.0
blis==0.4.1
boto==2.49.0
boto3==1.10.8
botocore==1.13.8
bottle==0.12.19
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
docutils==0.15.2
gast==0.2.2
gensim==3.8.1
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
numpy==1.17.3
opt-einsum==3.1.0
pandas==1.1.4
plac==1.1.3
preshed==3.0.2
protobuf==3.10.0
psycopg2-binary==2.8.4
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2020.4
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
smart-open==1.9.0
spacy==2.2.2
srsly==0.2.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
thinc==7.3.1
tqdm==4.37.0
urllib3==1.25.6
wasabi==0.3.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
